requirements.txt
requests
pandas
pandas-ta
numpy==1.26.4
fastapi
uvicorn
python-telegram-bot